Added to sphere_msgs.scp:    //menu_unexpected				"Unexpected menu info." Changed DWORD to unsigned long in CWorldClock::Advance() to avoid some "WARNING:system clock 0%lxh overflow - recycle\n" messages.
Now skill changes (on lumberjacking,tactics and anatomy when COMBATF_OSIDAMAGEMOD is enabled) and DAM value applied on weapons and the new CombatBonusStat and CombatBonusPercente are updating the status window to reflect the dam changes.
Fixed CombatBonusStat and CombatBonusPercent only working when COMBATF_OSIDAMAGEMOD was not enabled.

+++ 0.56.b/graysvr/CWorld.cpp
@@ -1048,9 +1048,9 @@ void CWorldClock::Init()
 bool CWorldClock::Advance()
 {
 	ADDTOCALLSTACK("CWorldClock::Advance");
-	DWORD Clock_Sys = GetSystemClock();	// get the system time.
+	unsigned long Clock_Sys = GetSystemClock();	// get the system time.
 
-	int iTimeSysDiff = Clock_Sys - m_Clock_PrevSys;
+	unsigned long iTimeSysDiff = Clock_Sys - m_Clock_PrevSys;
 	iTimeSysDiff = IMULDIVDOWN( TICK_PER_SEC, iTimeSysDiff, CLOCKS_PER_SEC );
 
 	if ( !iTimeSysDiff )
